# swagger/swagger.yaml
swagger: "2.0"
info:
  description: api user manager
  version: "1.0.0"
  title: Swagger REST API
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /customer/{id}:
    get:
      operationId: app.get_customer
      summary: 고객 정보 조회
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: 고객 ID
      responses:
        '200':
          description: '고객 정보 반환'
          schema:
            type: object
            properties:
              _id:
                type: string
                example: "66e14a6a272415eb9180ebbc"
              customer_name:
                type: string
                example: "홍길동"
        '404':
          description: '고객을 찾을 수 없음'

  # 고객 등록
...